# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Hifive football api",
    "version": "1.0.0",
    "title": "HifiveFootball",
    "termsOfService": "http://api.hifivefootball.com/",
    "contact": {
      "email": "ddastudio@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.hifivefootball.com:5300",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "get comments of content",
        "description": "get comments of content",
        "operationId": "getComments",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "contentid",
            "description": "content id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContentHeader"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "content",
        "description": "post content",
        "operationId": "postContent2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "post board",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCommentBody"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          },
          "400": {
            "description": "fail"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "comments"
        ],
        "summary": "content",
        "description": "post content",
        "operationId": "updateComment",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "post board",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCommentBody"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          },
          "400": {
            "description": "fail"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/comemnt/{id}": {
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "content",
        "description": "post content",
        "operationId": "deleteComment",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "id",
            "description": "commentid",
            "required": true,
            "type": "integer"
          },
          {
            "in": "formData",
            "name": "contentid",
            "description": "content_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "delete",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "400": {
            "description": "fail"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ContentHeader": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "arena_id": {
          "type": "integer"
        },
        "deleted": {
          "type": "integer"
        },
        "reported": {
          "type": "integer"
        },
        "comments": {
          "type": "integer"
        },
        "scraps": {
          "type": "integer"
        },
        "board_id": {
          "type": "integer"
        },
        "preview": {
          "type": "string"
        },
        "likers": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        },
        "bodytype": {
          "type": "integer"
        },
        "celltype": {
          "type": "integer"
        },        
        "link": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      },
      "xml": {
        "name": "ContentHeader"
      }
    },
...
    "Comment": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "group_id": {
          "type": "integer"
        },
        "depth": {
          "type": "integer"
        },
        "parent_id": {
          "type": "integer"
        },
        "content": {
          "type": "string"
        },
        "likers": {
          "type": "integer"
        },
        "deleted": {
          "type": "integer"
        },
        "reported": {
          "type": "integer"
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      },
      "xml": {
        "name": "Comment"
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserAction": {
      "type": "object",
      "properties": {
        "like": {
          "type": "integer"
        },
        "scrap": {
          "type": "integer"
        },
        "hasImgs": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "user_action"
      }
    },
...
    "PostCommentBody": {
      "type": "object",
      "properties": {
        "contentid": {
          "type": "integer"
        },
        "parentid": {
          "type": "integer"
        },
        "groupid": {
          "type": "integer"
        },
        "depth": {
          "type": "integer"
        },
        "content": {
          "type": "string"
        }
      },
      "xml": {
        "name": "PostContent"
      }
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  },
...
}